refactor: reorganize library exports
