+++ library/components/index.ts
@@ -1,0 +1,9 @@
+export { default as BlurOverlay } from './BlurOverlay.vue'
+export { default as LoadingOverlay } from './LoadingOverlay.vue'
+export { default as QrCode } from './QrCode.vue'
+export { default as Spinner } from './Spinner.vue'
+
+export type { props as BlurOverlayProps } from './BlurOverlay.vue'
+export type { props as LoadingOverlayProps } from './LoadingOverlay.vue'
+export type { props as QrCodeProps } from './QrCode.vue'
+export type { props as SpinnerProps } from './Spinner.vue'
